+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "sweetpad" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.9] — 2024-02-28
+
+- Detect schemes by reading files in workspace folder instead of using xcodebuild
+- Extract configuration name from .xcscheme file
+
 ## [0.0.8] - 2024-02-27
 
 - Add command to set active workspace